refactor: remove unit mapping from pogues variables
